feat: refactor ProjectsListener to dedicated project status listener
